third_party/blink/web_tests/http/tests/security/promise-access-control-deny-expected.txt

CONSOLE ERROR: Uncaught (in promise) 42
Tests that promise rejection events are muted for cross origin non-CORS scripts
